feat: MocknetworkManager 분리
#16
